... Detail Page

+++ src/app/houses/house-list/house-list.component.ts
@@ -45,12 +45,13 @@ export class HouseListComponent implements OnInit {
   }
 
   deleteHouse(id: number) {
-    console.log(id);
-    this.userService.deleteHouse(id, this.authService.decodedToken.nameid).subscribe(() => {
-      location.reload();
-    }, error => {
-      console.log(error);
-    });
+    if(confirm('Are you sure you want to delete?')) {
+      this.userService.deleteHouse(id, this.authService.decodedToken.nameid).subscribe(() => {
+        location.reload();
+      }, error => {
+        console.log(error);
+      });
+    }
   }
 
   updateUser(id: number) {
